requirements.txt
aioconsole==0.8.0
aiofiles==24.1.0
aiohttp==3.8.3
dnspython==2.7.0
humanfriendly==10.0
interactions.py==5.13.2
interactions_lavalink
lavalink
motor==3.4.0
Pillow==11.0.0
pymongo==4.8.0
PyYAML
deepl
python_dateutil==2.8.2
PyYAML==6.0.2
termcolor==2.5.0
humanize
aioconsole
termcolor
psutil
babel